Creating CreditCard ORM
