--- src/main/java/br/com/exactalabs/bicycleshop/repository/AdressRepository.java
@@ -1,13 +1,0 @@
-package br.com.exactalabs.bicycleshop.repository;
-
-import br.com.exactalabs.bicycleshop.entity.Adress;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AdressRepository extends JpaRepository<Adress, Long> {
-
-
-
-
-}
